custom response type

+++ HRLeaveManagementApplication/Features/LeaveAllocations/Handlers/Commands/DeleteLeaveAllocatoinCommandHandler.cs
@@ -1,6 +1,8 @@
 ﻿using AutoMapper;
+using HRLeaveManagementApplication.Exceptions;
 using HRLeaveManagementApplication.Features.LeaveAllocations.Requests.Commands;
 using HRLeaveManagementApplication.Persistence.Contracts;
+using HRLeaveManagementDomain;
 using MediatR;
 using System;
 using System.Collections.Generic;
@@ -22,6 +24,10 @@ namespace HRLeaveManagementApplication.Features.LeaveAllocations.Handlers.Comman
         public async Task<Unit> Handle(DeleteLeaveAllocationCommand request, CancellationToken cancellationToken)
         {
             var leaveAllocation = await _leaveAllocationRepository.Get(request.Id);
+            if (leaveAllocation == null)
+            {
+                throw new NotFoundException(nameof(leaveAllocation), request.Id);
+            }
             await _leaveAllocationRepository.Delete(leaveAllocation);
             return Unit.Value;
         }
